refactor: split disk health operations into dedicated usecase packages
