+++ internal/disk_health/usecases/operations/assess_dmesg/parser.go
@@ -1,4 +1,4 @@
-package usecases
+package assessdmesg
 
 import (
 	"fmt"
