+++ internal/disk_health/usecases/operations/assess_smart/assessor.go
@@ -1,4 +1,4 @@
-package usecases
+package assesssmart
 
 import (
 	"fmt"
